fix null warning

+++ KoiFishCare/Controllers/PrescriptionRecordController.cs
@@ -16,7 +16,7 @@ namespace KoiFishCare.Controllers
     [Route("api/pres-rec")]
     public class PrescriptionRecordController : ControllerBase
     {
-        
+
         private readonly IPrescriptionRecordRepository _preRecRepo;
         private readonly UserManager<User> _userManager;
         private readonly IBookingRepository _bookingRepo;
@@ -33,7 +33,7 @@ namespace KoiFishCare.Controllers
         public async Task<IActionResult> GetById([FromRoute] int bookingId)
         {
             var presRec = await _preRecRepo.GetById(bookingId);
-            if(presRec == null) return NotFound();
+            if (presRec == null) return NotFound();
             return Ok(presRec.ToPresRecDtoFromModel());
         }
 
@@ -42,7 +42,7 @@ namespace KoiFishCare.Controllers
         public async Task<IActionResult> GetListByCusname([FromRoute] string cusName)
         {
             var list = await _preRecRepo.GetListByCusUsername(cusName);
-            if(!list.Any()) return NotFound("There is no prescription record of this customer!");
+            if (!list.Any()) return NotFound("There is no prescription record of this customer!");
             var dto = list.Select(p => p.ToPresRecDtoFromModel());
             return Ok(dto);
         }
@@ -51,21 +51,22 @@ namespace KoiFishCare.Controllers
         [Authorize(Roles = "Staff, Vet")]
         public async Task<IActionResult> Create([FromBody] CreatePresRecordDto createDto)
         {
-            if(!ModelState.IsValid)
+            if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
             var booking = await _bookingRepo.GetBookingByIdAsync(createDto.BookingID);
-            if(booking == null) return NotFound("Not found booking!");
+            if (booking == null) return NotFound("Not found booking!");
 
             //check only the vet of the booking or staff can create
-            if(User.IsInRole("Vet"))
+            if (User.IsInRole("Vet"))
             {
                 var user = await _userManager.GetUserAsync(this.User);
-                if(user.Id != booking.VetID) return Unauthorized("This booking is not yours, you can not create record for this!");
+                if (user == null) return NotFound("Not found user");
+                if (user.Id != booking.VetID) return Unauthorized("This booking is not yours, you can not create record for this!");
             }
 
             var presRec = await _preRecRepo.Create(createDto.ToModelFromCreate());
-            if(presRec == null) return NotFound("Can not find BookingId");
+            if (presRec == null) return NotFound("Can not find BookingId");
 
             booking.hasPres = true;
             _bookingRepo.UpdateBooking(booking);
@@ -76,23 +77,25 @@ namespace KoiFishCare.Controllers
         [Authorize(Roles = "Staff, Vet")]
         public async Task<IActionResult> Update([FromRoute] int presRecId, [FromBody] UpdatePresRecordDto updateDto)
         {
-            if(!ModelState.IsValid)
+            if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
             var presReco = await _preRecRepo.GetById(presRecId);
-            if(presReco == null) return NotFound("Not found Pres Rec");
+            if (presReco == null) return NotFound("Not found Pres Rec");
             if (presReco.BookingID == null) return BadRequest("Booking ID is null");
             var booking = await _bookingRepo.GetBookingByIdAsync(presReco.BookingID.Value);
+            if (booking == null) return NotFound("Not found booking");
 
             //check only the vet of the booking or staff can create
-            if(User.IsInRole("Vet"))
+            if (User.IsInRole("Vet"))
             {
                 var user = await _userManager.GetUserAsync(this.User);
-                if(user.Id != booking.VetID) return Unauthorized("This booking is not yours, you can not update record for this!");
+                if (user == null) return NotFound("Not found user");
+                if (user.Id != booking.VetID) return Unauthorized("This booking is not yours, you can not update record for this!");
             }
 
             var presRec = await _preRecRepo.Update(presRecId, updateDto.ToModelFromUpdate());
-            if(presRec == null) return NotFound();
+            if (presRec == null) return NotFound();
 
             return Ok(presRec.ToPresRecDtoFromModel());
         }
@@ -102,18 +105,20 @@ namespace KoiFishCare.Controllers
         public async Task<IActionResult> Delete(int presRecId)
         {
             var preRec = await _preRecRepo.GetById(presRecId);
-            if(preRec == null) return NotFound("Not found Pres Rec");
+            if (preRec == null) return NotFound("Not found Pres Rec");
             if (preRec.BookingID == null) return BadRequest("Booking ID is null");
             var booking = await _bookingRepo.GetBookingByIdAsync(preRec.BookingID.Value);
+            if (booking == null) return NotFound("Not found booking");
 
-            if(User.IsInRole("Vet"))
+            if (User.IsInRole("Vet"))
             {
                 var user = await _userManager.GetUserAsync(this.User);
-                if(user.Id != booking.VetID) return Unauthorized("This booking is not yours, you can not delete record for this!");
+                if(user == null) return NotFound("Not found user");
+                if (user.Id != booking.VetID) return Unauthorized("This booking is not yours, you can not delete record for this!");
             }
 
             var presRec = await _preRecRepo.Delete(presRecId);
-            if(presRec == null) return NotFound();
+            if (presRec == null) return NotFound();
 
             return NoContent();
         }
